renamed base-return-data.ts to types.ts
