+++ src/test-linux.ts
@@ -68,7 +68,7 @@ export async function getApp(appName: string): Promise<InstalledApp | undefined>
 
 // Test the functions
 (async () => {
-  console.log('Testing isAppInstalled for "firefox":');
-  const result = await isAppInstalled('firefox');
+  console.log('Testing isAppInstalled for "activitywatch":');
+  const result = await isAppInstalled('activitywatch');
   console.log('Result:', result);
 })();
